Fix wavesystem this.wave being 0 every update
this.wave starts as 0, when an update runs, it would be set to 0 if the game mode is wave null.

When the run logic finishes, the wave gets incremented. However, when the update runs before the run code, the wave gets set back to 0. This update fixes that. In theory.